Add markdown rendering for all pings and metrics
