+++ src/components/MarkDescription.svelte
@@ -1,0 +1,7 @@
+<script>
+  import marked from "marked";
+
+  export let description;
+</script>
+
+{@html marked.parseInline(description)}
